Admin functionalities, isadmin added in delete and update functions, verify user checked, signup can be done as either admin or simple user

+++ api/controllers/user.controller.js
@@ -9,7 +9,7 @@ export const test = (req, res) => {
 };
 
 export const updateUser = async (req, res, next) => {
-  if (req.user.id !== req.params.id)
+  if (!(req.user.id === req.params.id || req.user.isAdmin))
     return next(errorHandler(401, 'You can only update your own account!'));
   try {
     if (req.body.password) {
@@ -38,9 +38,14 @@ export const updateUser = async (req, res, next) => {
 };
 
 export const deleteUser = async (req, res, next) => {
-  if (req.user.id !== req.params.id)
-    return next(errorHandler(401, 'You can only delete your own account!'));
   try {
+    const userToDelete = await User.findById(req.params.id);
+    
+    // Check if the user making the request is the user to be deleted or an admin
+    if (!(req.user.id === req.params.id || req.user.isAdmin)) {
+      return next(errorHandler(401, 'You are not authorized to delete this user!'));
+    }
+    
     await User.findByIdAndDelete(req.params.id);
     res.clearCookie('access_token');
     res.status(200).json('User has been deleted!');
